bueno, mejoramos el frontend y el backend en cuentas y jerarquia

+++ backend/src/main/java/com/hotel/backend/repository/ModeloAsiento_Repository.java
@@ -3,7 +3,7 @@ package com.hotel.backend.repository;
 import com.hotel.backend.entities.ModeloAsiento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModeloAsiento_Repository extends JpaRepository<ModeloAsiento, Long> {
+public interface ModeloAsiento_Repository extends JpaRepository<ModeloAsiento, Integer> {
 
 
 }
